+++ NEWS.md
@@ -1,5 +1,6 @@
 # bigrquerystorage 1.1.0.9000
 
+* Fix nested list parse post processing.
 * Fix returning more rows than actual rows in source table when n_max > nrows.
 
 # bigrquerystorage 1.1.0