+++ NEWS.md
@@ -1,5 +1,6 @@
 # bigrquerystorage 1.1.0.9000
 
+* Now uses nanoarrow instead of arrow. This lightens the dependency chain quite a bit (@hadleywickham).
 * With `use_tibble = TRUE`
   * Convert BigQuery SQL types BYTES, GEOGRAPHY to blob, wk_wkt
   * Set timezone of BigQuery SQL type DATETIME to UTC